More feedback:
- Separate business logic from event registration
- Consolidate clockUtils into utils
- Make util export functions instead of using class
